refactor: refactored TokenService and updated imports gf-8
