+++ apps/backend/.env.example
@@ -18,3 +18,4 @@ DB_POOL_MAX=10
 #
 JWT_SECRET=secret_key
 JWT_EXPIRATION_TIME=24h
+JWT_ALGORITHM="HS256"
