+++ apps/backend/src/modules/users/libs/types/types.ts
@@ -1,5 +1,6 @@
 export {
 	type UserGetAllResponseDto,
+	type UserResponseDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 } from "@git-fit/shared";
