+++ apps/backend/src/modules/users/user.service.ts
@@ -1,13 +1,11 @@
-import { type UserResponseDto } from "@git-fit/shared/src/modules/users/libs/types/types.js";
-
 import { type Service } from "~/libs/types/types.js";
-import { UserEntity } from "~/modules/users/user.entity.js";
-import { type UserRepository } from "~/modules/users/user.repository.js";
-
 import {
 	type UserGetAllResponseDto,
+	type UserResponseDto,
 	type UserSignUpRequestDto,
-} from "./libs/types/types.js";
+} from "~/modules/users/libs/types/types.js";
+import { UserEntity } from "~/modules/users/user.entity.js";
+import { type UserRepository } from "~/modules/users/user.repository.js";
 
 class UserService implements Service {
 	private userRepository: UserRepository;
