+++ packages/shared/src/modules/users/users.ts
@@ -2,6 +2,7 @@ export { UsersApiPath } from "./libs/enums/enums.js";
 export {
 	type UserGetAllItemResponseDto,
 	type UserGetAllResponseDto,
+	type UserResponseDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 } from "./libs/types/types.js";
